[solidity] bug fix

+++ regression/esbmc-solidity/dyn_array_7/contract.sol
@@ -3,15 +3,15 @@ pragma solidity >=0.5.0;
 
 // Return array of uint8 type
 contract A {
-    uint[2] z; // int *z = calloc(2,1);
+    uint8[2] z; // int *z = calloc(2,1);
     // int tmp1[2] = {1,2};
     // int *zz = calloc(2,1);
-    uint[2] zz = [1, 2]; // memcpy(zzzz, tmp2, 2);
+    uint8[2] zz = [1, 2]; // memcpy(zzzz, tmp2, 2);
 
-    uint[] zzz; // int* zzz = 0;
+    uint8[] zzz; // int* zzz = 0;
     // int tmp2[2] = {1,2};
     // int* zzzz = calloc(2,1);
-    uint[] zzzz = [1, 2]; // memcpy(zzzz, tmp2, 2);
+    uint8[] zzzz = [1, 2]; // memcpy(zzzz, tmp2, 2);
 
     function int8_call() public virtual returns (uint8[2] memory) {
         uint8[2] memory a; // int *a = calloc(2,1);
